refactor: Move collect_python_files to index/files module
- Extract file operations from index/mod.rs to dedicated files.rs module
- Prepare for future file watching functionality
- Keep file operations within the index module for better cohesion
